refactored data management to make things more modular
