+++ src/components/auth/Login.js
@@ -1,7 +1,7 @@
 import React, { useRef, useState } from "react";
 import { Link } from "react-router-dom";
 import { useHistory } from "react-router-dom";
-import { getUserByEmail } from "../ApiManager.js";
+import { getUserByEmail } from "../data_management/UsersData.js";
 import "./Login.css";
 
 export const Login = () => {
