+++ src/components/data_management/RudimentsData.js
@@ -1,0 +1,6 @@
+export const getAllRudiments = () => {
+    return fetch(`${API}/rudiments`).then((res) => res.json());
+};
+export const getRudimentById = (id) => {
+    return fetch(`${API}/rudiments/${id}`).then((res) => res.json());
+};
