+++ src/components/data_management/StudentsProfileData.js
@@ -1,0 +1,6 @@
+export const getStudentProfiles = () => {
+    return fetch(`${API}/studentsProfile`).then((res) => res.json());
+};
+export const postStudentProfile = (userProfile) => {
+    return fetch(`${API}/studentsProfile`, postOptions(userProfile));
+};
